+++ windows/src/developer/history.md
@@ -3,6 +3,9 @@
 ## 13.0 alpha
 * Start version 13.0
 
+## 2019-09-20 12.0.34 beta
+* Add basic wordlist editor to Keyman Developer (#2086)
+
 ## 2019-09-19 12.0.33 beta
 * Disable canonicalization of BCP 47 codes when building keyboard and model projects from template because it's inadequate (#2084)
 * Prevent package metadata preprocessor from treating .model.js files as keyboard files when they don't exist (#2085)